Topic/determinist

See merge request adrien_escande_at_gmail_com/VariableManagement!21

+++ src/hint/Substitutions.cpp
@@ -210,7 +210,7 @@ namespace internal
 
   void Substitutions::updateVariableValues() const
   {
-    for (int i = 0; i < variables_.size(); ++i)
+    for (size_t i = 0; i < variables_.size(); ++i)
     {
       varSubstitutions_[i]->updateValue();
       variables_[i]->value(varSubstitutions_[i]->value());
